20230831 TEST

+++ src/main/java/com/soccer/common/FileManagerService.java
@@ -20,7 +20,7 @@ public class FileManagerService {
 		private TextTransService textTransService;
 	
 	
-		public static final String FILE_UPLOAD_PATH = "E:\\7_FinalProject\\workspace\\images/";
+		public static final String FILE_UPLOAD_PATH = "D:\\choiminseok\\7_final_project\\workspace\\images/";
 		
 		
 		// 이미지 파일 저장 
@@ -53,15 +53,15 @@ public class FileManagerService {
 		// 이미지 파일 삭제 / 폴더 삭제
 		public void deleteFile(String profileImagePath) {
 			Path path = Paths.get(FILE_UPLOAD_PATH + profileImagePath.replace("/images/", ""));
-			
-			try {
-				Files.delete(path);
-			} catch (IOException e) {
-				logger.info("###[FileManagerService 이미지 삭제 실패] profileImagePath:{}", profileImagePath);
-			}
-			
-			// 디렉토리(폴터) 삭제
-			path = path.getParent();
+			if (Files.exists(path)) { // 이미지가 존재하는가?
+				try {
+					Files.delete(path);
+				} catch (IOException e) {
+					logger.info("###[FileManagerService 이미지 삭제 실패] profileImagePath:{}", profileImagePath);
+				}
+				
+				// 디렉토리(폴터) 삭제
+				path = path.getParent();
 				if (Files.exists(path)) {
 					try {
 						Files.delete(path);
@@ -69,8 +69,8 @@ public class FileManagerService {
 						logger.info("###[FileManagerService 폴더 삭제 실패] profileImagePath:{}", profileImagePath);
 					}
 				}
+			}
 		}
-
 }
